requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b
feedparser>=6.0.0
jinja2>=3.1.0
pyyaml>=6.0
pytest>=7.0.0
